+++ frontend/src/components/DownloadAnalysisLink/DownloadAnalysisLink.tsx
@@ -6,7 +6,7 @@ import { CONFIG } from '../../config/config'
 
 import { Store, withStore } from '../../services/store'
 import { download } from '../../utils/download'
-import { buildCsvFromData, getStaticCsvUrl } from './BuildAnalysis'
+import { buildCsvFromData, getStaticCsvUrl } from './BuildCSV'
 
 import './DownloadAnalysisLink.css'
